Add a x86-assembly implementation
The meat of this implementation is lifted completely from evanjones.ca. I
will follow up with Evan to check if that's OK (e.g. for licensing), and
will remove this if he's unhappy with it. (I wanted to have something to
point to before making contact.) His implementation is largely pedagogic,
and the upper levels of the code are not well-aligned with libfibre, but
his x86(-64) assembly tricks are *very* handy. I have taken cropings of
his code verbatim (though I made one change to suppress compiler warnings)
and put the libfibre stuff around that.

Full details (and acknowledgements and URLs) are in the code.

Note, some ad-hoc benchmarking on a 64-bit linux VM of the test_fibre
tool (with DO_TRACE undefined), gave approximate cycle counts for the
different arch implementations;
   ucontext: ~575 cycles per fibre switch
     setjmp: ~74 cycles per fibre switch
        x86: ~44 cycles per fibre switch

Signed-off-by: Geoff Thorpe <geoff@openssl.org>