Arbitrary changes to targets
If TARGET=debug-gcc, we default to ucontext if FIBRE_ARCH is undefined.
If TARGET=default-gcc (which is the default if TARGET is undefined), we
print an error if FIBRE_ARCH is undefined.

Long-term, the target stuff should be refactored, and actual high-level
targets defined that can steer all the lower-level definitions. (And
perhaps be dynamic/parameterisable for optimisation, debugging, ...)

Signed-off-by: Geoff Thorpe <geoff@openssl.org>