Add a setjmp/longjmp implementation
This follows some of the description given by Ralf Engelschall in his
paper on the subject. (It does not use GPL-tained code from GNU Pth
though.) The delta between Ralf's description and prior art seems to be
the realisation that a post-signal longjmp() pair is required to re-arm
the signal-generated fibre context, with a parameterless subroutine
required to avoid the "invalid stack frame" problems.

Oh, and I also found (and fixed) that _FORTIFY_SOURCE needs to be
undefined prior to including setjmp.h to avoid runtime "safety checks"
that break the longjmp()s even with Ralf's "clean stack frame" idea.
Removing those checks improve the speed noticably, as it happens.

Note: the use of pthread locking and sem_t semaphores requires linking
with "-lpthread", hence the Makefile.am update.

Signed-off-by: Geoff Thorpe <geoff@openssl.org>