test_fibre: change default behaviour
This routine is still compile-time controlled, I should find time to have
it use command-line options to determine behavior instead.

In any case, the current point of interest is now benchmarking, rather
than sanity-checking that fibres are getting created and switched
correctly. So DO_TRACE is undefined, and the loop limit is moved up to
100M iterations. (Much less than that, and the overheads of program
start-up and tear-down muddy the measurements. Especially with the setjmp
and x86 implementations.)

Signed-off-by: Geoff Thorpe <geoff@openssl.org>